Add findUserByEmail + rename findUser to findUserById
