+++ src/components/users/DALs/findUserByEmailDAL.ts
@@ -1,0 +1,5 @@
+import {User} from "..";
+
+export default async function(email: String) {
+  return await User.findOne({email});
+}
