+++ src/components/users/DALs/index.ts
@@ -1,4 +1,5 @@
-export {default as findUser} from './findUserDAL';
+export {default as findUserById} from './findUserById';
+export {default as findUserByEmail} from './findUserByEmailDAL';
 export {default as createUser} from './createUserDAL';
 export {default as addFridgeToUser} from './addFridgeToUserDAL';
 export {default as removeFridgeFromUser} from './removeFridgeFromUserDAL';
